backend/requirements.txt
# Core Framework
fastapi==0.121.2
pydantic==2.10.4
uvicorn==0.38.0

# Logging and Monitoring
structlog==25.5.0
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-fastapi==0.59b0
opentelemetry-instrumentation-httpx==0.59b0
opentelemetry-instrumentation-asyncpg==0.59b0

# Database and Storage
asyncpg==0.30.0
pymongo[srv]==4.6.1
motor==3.3.2


# AI/ML Dependencies
tiktoken==0.8.0
numpy
tqdm
google-generativeai==0.8.3
pinecone==7.3.0

# Environment and Configuration
# python-dotenv==1.0.1 # Moved to Utilities

# Utilities
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.0
slowapi==0.1.9

# Authentication
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
authlib==1.3.0

# HTTP Client (for health checks and external APIs)
httpx==0.28.1
requests==2.32.3

# Utilities
python-multipart==0.0.12
aiofiles==24.1.0

# Development and Testing (optional for production)
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
pytest-mock==3.14.0
black==24.10.0
flake8==7.1.1
mypy==1.14.1
